# docs/openapi/attendance.yaml
openapi: 3.0.3
info:
  title: HRM Attendance API
  version: 1.0.0
  description: Primary attendance endpoints and the temporary legacy compatibility routes.
servers:
  - url: http://167.172.95.86
security:
  - bearerAuth: []
paths:
...
  /employee/attendanceList/{cmpId}:
    get:
      operationId: getLegacyCompanyAttendance
      deprecated: true
      summary: Legacy company attendance list
      description: Use /api/attendance/company/{companyId}. Removal is no earlier than 15 calendar days after production rollout, after the full 14-day usage window; 2026-10-22 is provisional if rollout is 2026-10-07.
      parameters:
        - in: path
          name: cmpId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: Legacy response shape }
        "403": { description: Company access denied }
  /employee/lastattendanceList/{cmpId}:
    get:
      operationId: getLegacyLatestCompanyAttendance
      deprecated: true
      summary: Legacy latest company attendance list
      description: Use /api/attendance/company/{companyId}/latest. Removal is no earlier than 15 calendar days after production rollout, after the full 14-day usage window; 2026-10-22 is provisional if rollout is 2026-10-07.
      parameters:
        - in: path
          name: cmpId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: Legacy response shape }
        "403": { description: Company access denied }
  /employee/attendanceList/{cmpId}/{month}:
    get:
      operationId: getLegacyMonthlyCompanyAttendance
      deprecated: true
      summary: Legacy monthly company attendance list
      description: Use /api/attendance/company/{companyId}/month/{month}. Removal is no earlier than 15 calendar days after production rollout, after the full 14-day usage window; 2026-10-22 is provisional if rollout is 2026-10-07.
      parameters:
        - in: path
          name: cmpId
          required: true
          schema: { type: integer, format: int64 }
        - in: path
          name: month
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Legacy response shape }
        "403": { description: Company access denied }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...